backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
aiomysql==0.2.0
alembic==1.13.1
pymysql==1.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Config
python-decouple==3.8

# Validation
pydantic
pydantic-settings==2.2.1

# Utils
httpx==0.27.0